+++ Tests/Runtime/Unity.Properties.Tests/PropertyContainer/Value/PropertyContainerValueTests.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Collections.Generic;
-using JetBrains.Annotations;
 using NUnit.Framework;
 
 namespace Unity.Properties.Tests